refactor(vanilla): make comm depends on vanilla, not the other way around
reason is that this will make iteration of comm/storegen much easier at the cost of more coupling
