+++ comm/build.gradle.kts
@@ -3,6 +3,7 @@ dependencies {
     api("com.nukkitx.protocol:bedrock-v486:2.9.5-SNAPSHOT")
     // internal deps
     api(project(":game"))
+    api(project(":vanilla"))
     api(project(":threading"))
     implementation("com.github.bedrockecs:common")
 }
