+++ vanilla/build.gradle.kts
@@ -1,5 +1,4 @@
 dependencies {
-    api(project(":comm"))
     api(project(":game"))
     api(project(":vanilla:generated"))
     implementation("com.github.bedrockecs:common")
