+++ vanilla/src/main/kotlin/com/github/bedrockecs/vanilla/VanillaCoreConfiguration.kt
@@ -1,6 +1,5 @@
 package com.github.bedrockecs.vanilla
 
-import com.github.bedrockecs.vanilla.comm.VanillaCommConfiguration
 import com.github.bedrockecs.vanilla.data.VanillaBlockConfigurator
 import com.github.bedrockecs.vanilla.data.VanillaBlockSerialConfigurator
 import com.github.bedrockecs.vanilla.game.VanillaDimensionConfiguration
@@ -11,7 +10,6 @@ import org.springframework.context.annotation.Import
 @Configuration
 @Import(
     value = [
-        VanillaCommConfiguration::class,
         VanillaBlockConfigurator::class,
         VanillaBlockSerialConfigurator::class
     ]
