--- vanilla/src/main/kotlin/com/github/bedrockecs/vanilla/comm/VanillaCommConfiguration.kt
@@ -1,8 +1,0 @@
-package com.github.bedrockecs.vanilla.comm
-
-import org.springframework.context.annotation.ComponentScan
-import org.springframework.context.annotation.Configuration
-
-@Configuration
-@ComponentScan(basePackageClasses = [VanillaCommConfiguration::class])
-class VanillaCommConfiguration
